pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus"
version = "0.1.0"
description = "All things DiT"
readme = "README.md"
requires-python = ">=3.10"
dependencies = []

[project.optional-dependencies]
train = [
    "torch>=2.0",
    "diffusers>=0.37.0",
    "transformers>=4.41.2",
    "accelerate>=0.31.0",
    "peft>=0.11.1",
    "huggingface_hub",
    "mlflow>=2.0",
]
precomputed = [
    "nexus[train]",
    "mosaicml-streaming",
    "pyyaml",
    "safetensors",
]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration (slow, require GPU/model)",
]
testpaths = ["tests", "datasets/prepare/sstk/tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
